pyproject.toml
[project]
name = "imposting"
version = "0.1.0"
description = "ImPosting - Simple social media posts scheduler"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "Django==5.2",
    "django-cleanup==9.0.0",
    "django-browser-reload==1.18.0",
    "python-dotenv==1.1.0",
    "requests==2.32.3",
    "requests-cache==1.2.1",
    "requests-oauthlib==2.0.0",
    "whitenoise==6.9.0",
    "loguru==0.7.3",
    "social-auth-app-django==5.4.3",
    "pycryptodome>=3.22.0",
    "waitress==3.0.2",
    "pillow>=11.2.1",
    "ffmpeg-python>=0.2.0",
]
